refactor(compiler): 修复 clippy --all -D warnings 报出的全部 lint
- 多参数函数收拢为上下文结构体：IfBlocks、OwnerArena、ForwardPayloadInfo、
  BitmapContext/SwitchContext
- 删除并入 region_decisions 的 collect_allocs 死代码路径
- 取模比较改用 is_multiple_of、drain/iter().next() 改 mem::take/first_mut()、
  根槽解码与回写循环改用迭代器
- cargo fmt --all 归一格式；cargo nextest run --workspace 931 项全过

+++ crates/gugu-compiler/src/runtime/message.rs
@@ -976,7 +976,7 @@ impl ReturnNode {
             u64::from(batch.export_state)
                 | (u64::from(batch.observed) << 8)
                 | (u64::from(batch.capacity_class) << 16)
-                | (u64::from(batch.source.raw()) << 24),
+                | (batch.source.raw() << 24),
             Ordering::Relaxed,
         );
         self.integrity
